feat: allow float values for padding, radius commands
- accept float values in padding, radius commands,

- accept float values in main set-props function

+++ src/plugin/set-properties/utils/set-padding.ts
@@ -8,7 +8,7 @@ interface setPaddingProps {
 }
 
 export default function setPadding({ param, value, node }: setPaddingProps) {
-  const padding = parseInt(value);
+  const padding = parseFloat(value);
   const nodeTypeCheck = node.type === 'FRAME' && node.layoutMode !== 'NONE';
 
   if (nodeTypeCheck) {
